Move all page files to pages folder
